# packages/backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: TODOMaster API
  description: |
    TODOMaster是一个基于OKR（Objectives and Key Results）方法论的任务管理系统。
    本API提供了完整的用户管理、目标设置、关键结果跟踪和任务管理功能。
  version: 1.0.0
  contact:
    name: TODOMaster Team
    email: support@todomaster.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: 开发服务器
  - url: https://api.todomaster.com/v1
    description: 生产服务器
...
security:
  - bearerAuth: []
paths:
...
  /tasks:
    get:
      tags: [tasks]
      summary: 获取任务列表
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_progress, completed, cancelled]
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, urgent]
        - name: objectiveId
          in: query
          schema:
            type: string
            format: uuid
        - name: keyResultId
          in: query
          schema:
            type: string
            format: uuid
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: 任务列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TasksListResponse'

    post:
      tags: [tasks]
      summary: 创建任务
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDetailResponse'
  /tasks/{id}:
    get:
      tags: [tasks]
      summary: 获取任务详情
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 任务详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDetailResponse'

    put:
      tags: [tasks]
      summary: 更新任务
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: 更新成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskDetailResponse'

    delete:
      tags: [tasks]
      summary: 删除任务
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 删除成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
...
  /stats/tasks:
    get:
      tags: [stats]
      summary: 获取任务统计数据
      security:
        - bearerAuth: []
      parameters:
        - name: priority
          in: query
          schema:
            type: string
            enum: [low, medium, high, urgent]
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, in_progress, completed, cancelled]
      responses:
        '200':
          description: 任务统计数据
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskStatsResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        estimatedHours:
          type: number
        actualHours:
          type: number
        dueDate:
          type: string
          format: date
        completedAt:
          type: string
          format: date-time
        notes:
          type: string
        userId:
          type: string
          format: uuid
        objectiveId:
          type: string
          format: uuid
        keyResultId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        pages:
          type: integer

    # 请求模型
...
    CreateTaskRequest:
      type: object
      required:
        - title
        - objectiveId
      properties:
        title:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high, urgent]
          default: medium
        estimatedHours:
          type: number
        dueDate:
          type: string
          format: date
        objectiveId:
          type: string
          format: uuid
        keyResultId:
          type: string
          format: uuid
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, in_progress, completed, cancelled]
        priority:
          type: string
          enum: [low, medium, high, urgent]
        actualHours:
          type: number
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        message:
          type: string
...
    SuccessMessage:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
...
    TasksListResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                tasks:
                  type: array
                  items:
                    $ref: '#/components/schemas/Task'
                pagination:
                  $ref: '#/components/schemas/Pagination'
    TaskDetailResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                task:
                  $ref: '#/components/schemas/Task'
...
    TaskStatsResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                summary:
                  type: object
                statusDistribution:
                  type: object
                priorityDistribution:
                  type: object
                completionTrend:
                  type: array
...